separate units and layers
