+++ src/enzo/__init__.py
@@ -1,4 +1,5 @@
 from . import activation
 from . import layers
+from . import units
 
 __version__ = '0.0.1.dev3'
