+++ tests/test_units.py
@@ -4,8 +4,8 @@ import enzo
 
 
 def test_init_neuron():
-    n = enzo.layers.Neuron()
-    assert isinstance(n, enzo.layers.Neuron)
+    n = enzo.units.Neuron()
+    assert isinstance(n, enzo.units.Neuron)
     assert n.weights is None
     assert n.bias is None
     assert n.activation is enzo.activation.noactivation
@@ -13,7 +13,7 @@ def test_init_neuron():
 
 def test_neuron_process_given_weights():
     inp = [0.2, 99, -5, 12]
-    n = enzo.layers.Neuron([[4.1], [0], [-2], [5]], bias=4)
+    n = enzo.units.Neuron([[4.1], [0], [-2], [5]], bias=4)
     assert n.process(inp) == 74.82
 
 
@@ -25,9 +25,9 @@ def patched_randn(*args):
 
 
 def test_neuron_process_without_weights(mocker):
-    patched = mocker.patch('enzo.layers.np.random.randn',
+    patched = mocker.patch('enzo.units.np.random.randn',
                            side_effect=patched_randn)
     inp = [0.2, 99, -5, 12]
-    n = enzo.layers.Neuron()
+    n = enzo.units.Neuron()
     assert n.process(inp) == pytest.approx(-19.85807)
     assert patched.call_args_list == [((1, 4),), ()]
